gateway/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.0

# JWT and authentication
python-jose[cryptography]==3.3.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Supabase client
supabase==2.10.0

# HTTP client
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1
